crawler/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
supabase==2.25.0
python-dotenv==1.0.1
websockets>=15.0.0
